Renamed widget functions to conform to standard
